Refactor component nesting
