+++ src/components/cart/CartTable.js
@@ -1,0 +1,15 @@
+import React from 'react';
+
+const CartTable = ({ children }) => {
+  return (
+    <>
+      <div className="col_table_wrap order_list">
+        {children}
+      </div>
+    </>
+  );
+};
+
+export default CartTable;
+
+
